Complete Feature Commit
this commit adds a player capsule at 0,1,0 scaled at 1,1,1

this commit will be pushed to the feature branch and a pull request issued immediately. To demonstrate how SmartMerge will save us, the pull request won't be approved until the feature in feature branch 2 is also pushed to that branch and the pull request is issued. This matches a pattern of managing pull requests at standard time(s) in sprints/weeks