create and add material. Complete issue
showing commit 2 in feature branch. This issue is now done and the pull request will be issued. At this point, the pull request from feature1 is still waiting to be approved and is not yet merged to develop.

Following this, pull request 1 will be approved and then pull request 2. This would normally cause a merge conflict.